Fixed cross-reference, separated state stores for Toolbar buttons
