+++ src/renderer/runtime/operations.ts
@@ -1,6 +1,6 @@
 import { Analytics } from "./analytics";
-import { appClipboard, ClipboardData, ClipboardKey } from "./clipboard.store";
-import { logger, runtime, user_input } from "./runtime.store";
+import { appClipboard, ClipboardKey } from "./clipboard.store";
+import { logger, user_input } from "./runtime.store";
 import {
   GridOperationResult,
   ElementData,
@@ -16,8 +16,7 @@ import {
   GridModule,
   GridPresetData,
 } from "./runtime";
-import { get, derived } from "svelte/store";
-import { config_panel_blocks } from "../main/panels/configuration/Configuration";
+import { get } from "svelte/store";
 
 function handleError(e: GridOperationResult) {
   //TODO: Better error handling
@@ -50,16 +49,6 @@ function handleError(e: GridOperationResult) {
 }
 
 //Clipboard handlers
-export const isCopyElementEnabled = derived(
-  config_panel_blocks,
-  ($config_panel_blocks) => {
-    return (
-      !$config_panel_blocks.some((e) => e.selected) &&
-      runtime.modules.length > 0
-    );
-  }
-);
-
 export async function copyElement(element: GridElement) {
   logger.set({
     type: "progress",
@@ -88,13 +77,6 @@ export async function copyElement(element: GridElement) {
     });
 }
 
-export const isCopyActionsEnabled = derived(
-  config_panel_blocks,
-  ($config_panel_blocks) => {
-    return $config_panel_blocks.some((e) => e.selected);
-  }
-);
-
 export async function copyActions(...actions: GridAction[]) {
   appClipboard
     .copyActions(...actions)
@@ -117,19 +99,6 @@ export async function copyActions(...actions: GridAction[]) {
 }
 
 //GridElement handlers
-export function isOverwriteElementEnabled(
-  data: ElementData,
-  clipboard: ClipboardData
-) {
-  if (typeof clipboard === "undefined" || typeof data === "undefined") {
-    return false;
-  }
-
-  return (
-    clipboard.key === ClipboardKey.ELEMENT &&
-    data.isCompatible((clipboard.payload as ElementData).type)
-  );
-}
 
 export async function overwriteElement(target: GridElement) {
   const clipboard = get(appClipboard);
@@ -159,14 +128,6 @@ export async function overwriteElement(target: GridElement) {
     });
 }
 
-export function isDiscardElementEnabled(data: ElementData) {
-  if (typeof data === "undefined") {
-    return false;
-  }
-
-  return data.hasChanges();
-}
-
 export async function discardElement(target: GridElement) {
   logger.set({
     type: "progress",
@@ -194,10 +155,6 @@ export async function discardElement(target: GridElement) {
         mandatory: false,
       });
     });
-}
-
-export function isClearElementEnabled(data: ElementData) {
-  return typeof data !== "undefined";
 }
 
 export async function clearElement(target: GridElement) {
@@ -264,13 +221,6 @@ export async function updateAction(
 
 //GridEvent handlers
 
-export const isMergeActionsEnabled = derived(
-  config_panel_blocks,
-  ($config_panel_blocks) => {
-    return $config_panel_blocks.some((e) => e.selected);
-  }
-);
-
 export async function mergeActionsToCode(
   target: GridEvent,
   ...actions: GridAction[]
@@ -283,10 +233,6 @@ export async function mergeActionsToCode(
     .catch(handleError)
     .finally(() => {});
 }
-
-export const isPasteActionsEnabled = derived(appClipboard, ($appClipboard) => {
-  return $appClipboard?.key === ClipboardKey.ACTION_BLOCKS;
-});
 
 export async function pasteActions(target: GridEvent, index?: number) {
   target
@@ -303,13 +249,6 @@ export async function pasteActions(target: GridEvent, index?: number) {
       });
     });
 }
-
-export const isRemoveActionsEnabled = derived(
-  config_panel_blocks,
-  ($config_panel_blocks) => {
-    return $config_panel_blocks.some((e) => e.selected);
-  }
-);
 
 export async function removeActions(
   target: GridEvent,
@@ -329,13 +268,6 @@ export async function removeActions(
       });
     });
 }
-
-export const isCutActionsEnabled = derived(
-  config_panel_blocks,
-  ($config_panel_blocks) => {
-    return $config_panel_blocks.some((e) => e.selected);
-  }
-);
 
 export async function cutActions(target: GridEvent, ...actions: GridAction[]) {
   appClipboard
